Removed unnecessary friendships in blogservice
