+++ backend/BlogService/Helpers/MappingProfile.cs
@@ -25,9 +25,6 @@ public class MappingProfile : Profile
         CreateMap<SavedBlog, SavedBlogDto>();
         CreateMap<CreateSavedBlogDto, SavedBlog>();
         
-        CreateMap<Friendship, FriendshipDto>()
-            .ForMember(dest => dest.Status, opt => opt.MapFrom(src => src.Status.ToString()));
-        
         CreateMap<Report, ReportDto>();
         CreateMap<CreateReportDto, Report>();
     }
