--- backend/BlogService/Models/Entities/FriendshipStatus.cs
@@ -1,9 +1,0 @@
-namespace BlogService.Models.Entities;
-
-public enum FriendshipStatus
-{
-    Pending = 0,
-    Accepted = 1,
-    Rejected = 2,
-    Blocked = 3
-}
